+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+###Added
+- Added Group post scheduling [#4027](https://github.com/rokwire/illinois-app/issues/4027).
 ### Changed
 - Require privacy level 4 for ability to create groups [#4030](https://github.com/rokwire/illinois-app/issues/4030).